@@ -22,21 +22,21 @@
     "@colyseus/loadtest": "^0.16.0",
     "@colyseus/testing": "^0.16.0",
     "@types/cors": "^2.8.6",
-    "@types/express": "^4.17.1",
-    "@types/mocha": "^8.2.3",
+    "@types/express": "^5.0.6",
+    "@types/mocha": "^10.0.10",
     "copyfiles": "^2.4.1",
-    "mocha": "^9.0.2",
-    "rimraf": "^2.7.1",
+    "mocha": "^11.7.5",
+    "rimraf": "^6.1.2",
     "ts-node": "^10.9.0",
     "ts-node-dev": "^2.0.0",
-    "typescript": "^4.9.5"
+    "typescript": "^5.9.3"
   },
   "dependencies": {
     "@colyseus/monitor": "^0.16.0",
-    "@colyseus/tools": "^0.16.0",
     "@colyseus/schema": "^3.0.0",
+    "@colyseus/tools": "^0.16.19",
     "colyseus": "^0.16.0",
     "cors": "^2.8.5",
-    "express": "^4.18.1"
+    "express": "^5.2.1"
   }
 }
